+++ aiohttp_middlewares/shield.py
@@ -42,18 +42,18 @@ import asyncio
 import logging
 
 from aiohttp import web
+from aiohttp.web_middlewares import _Handler, _Middleware
 
-from .annotations import Handler, Middleware, StrCollection, Urls
+from .annotations import StrCollection, Urls
 from .utils import match_request
 
 
 logger = logging.getLogger(__name__)
 
 
-def shield_middleware(*,
-                      methods: StrCollection = None,
-                      urls: Urls = None,
-                      ignore: Urls = None) -> Middleware:
+def shield_middleware(
+    *, methods: StrCollection = None, urls: Urls = None, ignore: Urls = None
+) -> _Middleware:
     """Ensure that handler execution would not break on ``CancelledError``.
 
     Shielding handlers allow to avoid breaking handler execution on
@@ -103,49 +103,48 @@ def shield_middleware(*,
         regexps from shielding. Do not mix with ``urls``.
     """
     if not methods and not urls:
-        raise ValueError('None of methods or urls argument passed.')
+        raise ValueError("None of methods or urls argument passed.")
     if methods and urls:
         raise ValueError(
-            'Both methods and urls arguments passed, while only one expected.')
+            "Both methods and urls arguments passed, while only one expected."
+        )
     if urls and ignore:
-        raise ValueError('Unable to mix urls and ignore arguments.')
+        raise ValueError("Unable to mix urls and ignore arguments.")
 
     # Lower case methods to shield (if any)
     methods_to_shield = {item.lower() for item in methods or []}
 
-    async def factory(app: web.Application, handler: Handler) -> Handler:
-        """Actual shield middleware factory."""
-        async def middleware(request: web.Request) -> web.Response:
-            """Shield handler execution if necessary."""
-            request_method = request.method.lower()
-            request_path = request.rel_url.path
-            log_extra = {'method': request.method, 'path': request_path}
+    @web.middleware
+    async def middleware(
+        request: web.Request, handler: _Handler
+    ) -> web.StreamResponse:
+        """Shield handler execution if necessary."""
+        request_method = request.method.lower()
+        request_path = request.rel_url.path
+        log_extra = {"method": request.method, "path": request_path}
 
-            # First attempt to process methods to shield
-            if methods_to_shield:
-                if request_method not in methods_to_shield:
-                    return await handler(request)
+        # First attempt to process methods to shield
+        if methods_to_shield:
+            if request_method not in methods_to_shield:
+                return await handler(request)
 
-                if (
-                    ignore and
-                    match_request(ignore, request_method, request_path)
-                ):
-                    logger.debug(
-                        'Ignore path from handler shielding.',
-                        extra=log_extra)
-                    return await handler(request)
-
+            if ignore and match_request(ignore, request_method, request_path):
                 logger.debug(
-                    'Activate shield middleware by matched method',
-                    extra=log_extra)
-                return await asyncio.shield(handler(request))
+                    "Ignore path from handler shielding.", extra=log_extra
+                )
+                return await handler(request)
 
-            # Then attempt to shield handler by URLs collection / mapping
-            if urls and match_request(urls, request_method, request_path):
-                logger.debug(
-                    'Activate shield middleware by matched path',
-                    extra=log_extra)
-                return await asyncio.shield(handler(request))
-            return await handler(request)
-        return middleware
-    return factory
+            logger.debug(
+                "Activate shield middleware by matched method", extra=log_extra
+            )
+            return await asyncio.shield(handler(request))
+
+        # Then attempt to shield handler by URLs collection / mapping
+        if urls and match_request(urls, request_method, request_path):
+            logger.debug(
+                "Activate shield middleware by matched path", extra=log_extra
+            )
+            return await asyncio.shield(handler(request))
+        return await handler(request)
+
+    return middleware